requirements.txt
Brotli==1.0.9
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
colorama==0.4.4
cycler==0.10.0
dash==2.0.0
dash-bootstrap-components==0.13.1
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
Flask==2.0.1
Flask-Compress==1.10.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.4.3
mysql-connector==2.2.9
numpy==1.21.2
pandas==1.3.3
pathlib==1.0.1
Pillow==8.3.2
plotly==5.3.1
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
scikit-learn==1.0
scipy==1.7.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
tenacity==8.0.1
threadpoolctl==2.2.0
urllib3==1.26.7
Werkzeug==2.0.1
